requirements.txt
# Selenium for web automation
selenium==4.10.0

# MongoDB driver for NoSQL database
pymongo==4.5.0

# Google Sheets API integration
gspread==5.9.0
oauth2client==4.1.3

# SQLite (optional, included in Python standard library but useful for completeness)
sqlite-utils==3.31.1

# dotenv for loading environment variables
python-dotenv==1.0.0
